% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anidb-package.R
\docType{package}
\name{anidb-package}
\alias{anidb}
\alias{anidb-package}
\title{anidb: An UDP client for AniDB}
\description{
AniDB (https://anidb.net/) is anime database that contains detailed information about Chinese, Japanese and Korean animation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emitanaka/anidb}
  \item Report bugs at \url{https://github.com/emitanaka/anidb/issues}
}

}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID})

}
\keyword{internal}
