% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{anidb_authenticate_user}
\alias{anidb_authenticate_user}
\title{Authenticate AniDB user}
\usage{
anidb_authenticate_user(
  user = NULL,
  password = askpass::askpass("user password")
)
}
\arguments{
\item{user}{The username for AniDB}

\item{password}{The password used for AniDB}
}
\description{
Authenticate AniDB user
}
