% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{anime_info}
\alias{anime_info}
\title{Attributes of anime}
\usage{
anime_info(aid = NULL)
}
\arguments{
\item{aid}{The vector of anime aid.}
}
\value{
Return a data frame that contains information about anime.
}
\description{
Attributes of anime
}
