% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{anime_ratings_dist}
\alias{anime_ratings_dist}
\title{Get the rating distribution for an anime}
\usage{
anime_ratings_dist(aid = NULL, type = c("permanent", "temporary"))
}
\arguments{
\item{aid}{The anime id}

\item{type}{A single string \code{"permanent"} or \code{"temporary"} to indicate the
type of rating distribution.}
}
\description{
Get the rating distribution for an anime
}
