% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{anime_ratings_gender_dist}
\alias{anime_ratings_gender_dist}
\title{Get the rating distribution for an anime}
\usage{
anime_ratings_gender_dist(
  aid = NULL,
  type = c("raw", "weighted"),
  include_title = TRUE
)
}
\arguments{
\item{aid}{The vector of anime aid.}

\item{type}{Whether to get the raw average ratings or the weighted average ratings.}

\item{include_title}{A logical value indicating whether to include the title in the
output data frame.}
}
\description{
Get the rating distribution for an anime
}
