% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{anime_ratings_ts}
\alias{anime_ratings_ts}
\title{Get the time series for the ratings}
\usage{
anime_ratings_ts(aid = NULL)
}
\arguments{
\item{aid}{The anime aid}
}
\description{
Get the time series for the ratings
}
