% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titles.R
\docType{data}
\name{animetitles}
\alias{animetitles}
\title{The full anime titles database}
\format{
A data frame that contains. The data frame also contains the attribute \code{date} which is
the date that the database was downloaded (\code{attr(animetitles, "date")}).
\describe{
\item{aid}{Anime ID in AniDB}
\item{type}{The title type: "short-title", "official", "primary", and "synonym".
There is only one primary title per anime ID. Primary titles are written
in Latin.}
\item{language}{The language of the title. The Latin version of the title
has the prefix "x-" and suffix "t", e.g. "x-jat" means that the corresponding
\code{title} is the Latin version of Japanese title.}
\item{title}{The anime title.}
}
}
\source{
\url{https://anidb.net/}
}
\usage{
animetitles
}
\description{
The full anime titles database
}
\keyword{datasets}
