% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{official_title}
\alias{official_title}
\title{Get the official anime title}
\usage{
official_title(aid, type = c("primary", "english", "original"))
}
\arguments{
\item{aid}{The anime ID.}

\item{type}{The title type: "primary", "english" (if available) or "original".}
}
\description{
Get the official anime title
}
