% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titles.R
\docType{data}
\name{officialtitles}
\alias{officialtitles}
\title{The official anime titles}
\format{
A data frame that contains the official anime titles and corresponding
anime ID in AniDB. The data frame also contains the attribute \code{date} which is
the date that the database was downloaded (\code{attr(officialtitles, "date")}).
\describe{
\item{aid}{Anime ID in AniDB}
\item{origin}{The country of origin. There are only three: \code{JP} = Japan, \code{CN} = China, \code{KR} = Korea.}
\item{title_primary}{The primary anime title. Primary titles are generally
the Latin form of the original titles.}
\item{title_en}{Official title in English (if available).}
\item{title_original}{The official title in the country of origin.}
}
}
\source{
\url{https://anidb.net/}
}
\usage{
officialtitles
}
\description{
The official anime titles
}
\keyword{datasets}
