% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udp.R
\name{udp_send}
\alias{udp_send}
\title{UDP send}
\usage{
udp_send(message, host = "localhost", port)
}
\arguments{
\item{message}{The messaage}

\item{host}{The host name.}

\item{port}{The port number}
}
\description{
UDP send
}
